+++ test/go.mod
@@ -1,8 +1,9 @@
-module github.com/w3f/terraform-google-polkadot-deployer
+module github.com/vladimir-babichev/terraform-google-polkadot-deployer/test
 
 go 1.14
 
 require (
 	github.com/gruntwork-io/terratest v0.27.2
 	github.com/stretchr/testify v1.5.1
+	github.com/vladimir-babichev/terratest-polkadot-deployer v0.0.0-20200514000955-c9fbb12dbeb5
 )
